Get ready for `@webref/css` v7 release, drop `alpha` flag
Time to give v7 a try in real life, following iterations and discussions in:
https://github.com/w3c/webref/issues/1519

Note: the way we managed alpha versions was somewhat clunky. The version
increment should have appeared *after* the `alpha` flag. This update drops the
`alpha` flag, which means that the first actual published package will be
`7.0.11` and not `7.0.0`.

Plan is to continue to publish v6 for a little while to give consumers enough
time to adjust their code and provide a fallback plan should the new version
turn out to be impractical.