Make NPM_TOKEN optional for package release (#1744)
Classic tokens are no longer supported by npm. We may still want to run the
release script from a local machine using a fine-grained access token, but
these tokens expire after 90 days at most and are thus not suitable for our
release process.

I set up OpenID Connect between the `@webref/*` packages in npm and GitHub
Actions and dropped the former `NPM_TOKEN` secret. This update adjusts the
release script not to fail if such a token cannot be found. The call to
`npmPublish` gets adjusted accordingly only to pass the token if it exists.

That should close #1739.